cpyext: check the exported declarations against CPython's
`scripts/cpyext-abi.py check` reads every `#[unsafe(no_mangle)] extern "C"`
function under `cpyext` and compares it to the declaration CPython
publishes for the same name, recorded in `scripts/cpython-abi.txt`. It
compares ABI slots rather than spellings, resolving CPython's typedefs, so
`Py_hash_t` and `Py_ssize_t` agree, a `PyCapsule_Destructor` and a `void *`
agree, and `long` and `Py_ssize_t` do not.

The record is a file because CI has no CPython checkout. `snapshot`
rewrites it from one; that diff is the ABI change a version bump makes.

Nothing else in this tree can see this class of failure: every fixture is
compiled against pyre's own header, so a parameter declared at the wrong
width agrees with itself. `PyModule_AddIntConstant` took a `Py_ssize_t`
where the reference declaration says `long` -- the same type on LP64 and
32 bits against 64 on Windows. It now takes a `c_long`, and
`PyModule_AddIntMacro` drops the cast that hid the difference.

327 of the 362 exports have a CPython declaration and all 327 match. The
other 35 are the macros CPython spells over struct fields, which have to
be calls here, plus two `_PyPyre_*` helpers.

+++ pyre/pyre-interpreter/src/cpyext/modsupport.rs
@@ -5,7 +5,7 @@ use super::methodobject::{CPyMethodDef, METH_CLASS, METH_COEXIST, METH_STATIC, n
 use super::pyerrors::{self, trap};
 use super::pyobject::{self, CPyObject, REFCNT_IMMORTAL};
 use pyre_object::{PY_NULL, PyObjectRef};
-use std::ffi::{CStr, c_char, c_int, c_void};
+use std::ffi::{CStr, c_char, c_int, c_long, c_void};
 use std::sync::atomic::{AtomicIsize, Ordering};
 
 /// `Py_mod_create`.
@@ -518,11 +518,13 @@ pub unsafe extern "C" fn PyModule_AddObjectRef(
     0
 }
 
+/// The value is a C `long`, which is the width the reference declaration says
+/// and narrower than a `Py_ssize_t` wherever the two differ.
 #[unsafe(no_mangle)]
 pub unsafe extern "C" fn PyModule_AddIntConstant(
     module: *mut CPyObject,
     name: *const c_char,
-    value: isize,
+    value: c_long,
 ) -> c_int {
     let Some(module) = module_argument(module, "PyModule_AddIntConstant") else {
         return -1;
